optional constexpr sq-antipr. angle lookup table
Adds the compile-time option to alter square antiprismatic's angle
function into a table lookup whose table is generated via constexpr
from its coordinates.

This is really only reasonable for that exact symmetry. For all other
symmetries, the approximate angle functions that group angles together
and compute only on the passed indices are good approximations. For
square-antiprismatic, this isn't really the case. A lookup table is a
reasonable alternative that is more correct, faster and not too wasteful.