Made minAngle calculation constexpr
This requires one to split SymmetryInformation into parts that are
literal types (and can be marked constexpr) and ones that can't.
Rotations, tetrahedra and coordinate lists cannot be constexpr. They all
depend on size for their respective array sizes, have arbitrary second
dimension lengths or require non-constexpr computations.

It's pretty ugly overall. In a new namespace AngleFunctions, all angle
functions are extruded. Then make a 1 to 1 sequence-matching array with
allNames containing size and angleFunction pointers. From that
calculating the smallest returned angle is simple.

Maybe it would be for the best to instead just set the minimum angle as
constant from the analysis/tests and check if it's still correct for
every test run...